+++ dispatcher.go
@@ -103,9 +103,10 @@ func addObservingMiddleware(cfg Config, meter *metrics.Scope, logger *zap.Logger
 	}
 
 	observer := observability.NewMiddleware(observability.Config{
-		Logger:           logger,
-		Scope:            meter,
-		ContextExtractor: extractor,
+		Logger:              logger,
+		Scope:               meter,
+		ContextExtractor:    extractor,
+		MetricTagsBlocklist: cfg.Metrics.TagsBlocklist,
 		Levels: observability.LevelsConfig{
 			Default: observability.DirectionalLevelsConfig{
 				Success:          cfg.Logging.Levels.Success,